Move all classes into separate files
