+++ src/AnimatedTransform.js
@@ -1,0 +1,78 @@
+/**
+ * Copyright (c) 2015-present, Facebook, Inc.
+ * All rights reserved.
+ *
+ * This source code is licensed under the BSD-style license found in the
+ * LICENSE file in the root directory of this source tree. An additional grant
+ * of patent rights can be found in the PATENTS file in the same directory.
+ *
+ * @flow
+ */
+'use strict';
+
+var Animated = require('./Animated');
+var AnimatedWithChildren = require('./AnimatedWithChildren');
+
+class AnimatedTransform extends AnimatedWithChildren {
+  _transforms: Array<Object>;
+
+  constructor(transforms: Array<Object>) {
+    super();
+    this._transforms = transforms;
+  }
+
+  __getValue(): Array<Object> {
+    return this._transforms.map(transform => {
+      var result = {};
+      for (var key in transform) {
+        var value = transform[key];
+        if (value instanceof Animated) {
+          result[key] = value.__getValue();
+        } else {
+          result[key] = value;
+        }
+      }
+      return result;
+    });
+  }
+
+  __getAnimatedValue(): Array<Object> {
+    return this._transforms.map(transform => {
+      var result = {};
+      for (var key in transform) {
+        var value = transform[key];
+        if (value instanceof Animated) {
+          result[key] = value.__getAnimatedValue();
+        } else {
+          // All transform components needed to recompose matrix
+          result[key] = value;
+        }
+      }
+      return result;
+    });
+  }
+
+  __attach(): void {
+    this._transforms.forEach(transform => {
+      for (var key in transform) {
+        var value = transform[key];
+        if (value instanceof Animated) {
+          value.__addChild(this);
+        }
+      }
+    });
+  }
+
+  __detach(): void {
+    this._transforms.forEach(transform => {
+      for (var key in transform) {
+        var value = transform[key];
+        if (value instanceof Animated) {
+          value.__removeChild(this);
+        }
+      }
+    });
+  }
+}
+
+module.exports = AnimatedTransform;
